pyproject.toml
[dependency-groups]
dev = [
    "antiberty>=0.1.3",
    "better-abc>=0.0.3",
    "biopython>=1.86",
    "ipykernel>=7.1.0",
    "joblib>=1.5.3",
    "jupyter>=1.1.1",
    "lab>=8.8",
    "lightning>=2.6.1",
    "llvmlite>=0.40.0",
    "logging>=0.4.9.6",
    "logomaker>=0.8.7",
    "matplotlib>=3.10.8",
    "mosaicml-streaming>=0.10.0",
    "numba>=0.57.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "peptides>=0.5.0",
    "python-dotenv>=1.2.1",
    "pytorch-lightning>=2.6.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers<4.40.0",
    "umap-learn>=0.5.11",
    "wandb>=0.25.0",
    "xgboost>=3.2.0",
]
